multiple pyomo GPs working and tested with ipopt
