+++ examples/pyomo_maximize/pyomo_maximize.py
@@ -60,38 +60,46 @@ def main():
 
     # Convert FoKL's GP model to Pyomo:
     scenarios = f.draws  # Pyomo 'scenarios' is synonym for FoKL 'draws'
-    m = f.to_pyomo(draws=scenarios)  # default is 'draws=self.draws' so this is unnecessary but demonstrates idea
+    xvars = ['x0', 'x1']
+    m = f.to_pyomo(xvars, 'y', draws=scenarios)  # default is 'draws=self.draws' so this is unnecessary but demonstrates idea
 
-    # Add constraints (if any/known) to enforce 'physics':
+    # Add known constraints (if any) to enforce 'physics':
     m.t = pyo.Var(within=pyo.Reals, bounds=[0, 1])
-    m.x0 = pyo.Constraint(expr=m.fokl_x[0] == (sin(m.t * np.pi * 3) + 1) / 2)
-    m.x1 = pyo.Constraint(expr=m.fokl_x[1] == (cos(m.t * np.pi * 3) + 1) / 2)
+    m.known = ConstraintList()
+    m.known.add(m.x0 == sin(m.t * np.pi * 3))
+    m.known.add(m.x1 == cos(m.t * np.pi * 3))
 
-    m.fokl_y_avg = pyo.Expression(expr=sum(m.fokl_y[i] for i in m.fokl_scenarios) / scenarios)  # average of scenarios
-    m.obj = pyo.Objective(expr=m.fokl_y_avg, sense=pyo.maximize)
+    # Set objective:
+    m.obj = pyo.Objective(expr=m.y, sense=pyo.maximize)
 
     # Solve (need multistart 'https://pyomo.readthedocs.io/en/latest/contributed_packages/multistart.html'):
-    # solver = pyo.SolverFactory('multistart')  # multistart to try local solver at various points for global solution
+    solver = pyo.SolverFactory('multistart')  # multistart to try local solver at various points for global solution
     print("\nRunning Pyomo solver...")
-    # solver.solve(m, solver='ipopt', suppress_unbounded_warning=True)  # IPOPT is local solver, and warning not needed
-    # print("Done!")
-
-    solver = pyo.SolverFactory('ipopt')
-    solver.solve(m, tee=True)
+    solver.solve(m, solver='ipopt', suppress_unbounded_warning=True)  # IPOPT is local solver, and warning not needed
+    print("Done!")
 
     print("\nPyomo solution:")
     print(f"     y = {m.obj()}")
     x = []
-    for j in m.fokl_j:
-        print(f"    x{j} = {m.fokl_x[j]()}")
-        x.append(m.fokl_x[j]())
+    x_norm = []
+    for j in m.GP0_j:  # = range(f.inputs.shape[1]) = range(len(xvars))
+        print(f"    x{j} = {m.component(xvars[j])()}")
+        x.append(m.component(xvars[j])())
+        x_norm.append(m.component(f"GP0_{xvars[j]}_norm")())
     print(f"     t = {m.t()}")
 
     # Check solution against FoKL model:
-    x = np.array(x)[np.newaxis, :]
+
+    print("\nNumerical FoKL evaluation of Pyomo solution [x0, x1] to confirm equivalence of models from...")
+
+    x = f.clean(x, SingleInstance=True)  # automatically format and normalize the true scale solution
     y_check = f.evaluate(x, draws=scenarios)
-    print("\nNumerical FoKL evaluation of Pyomo solution [x0, x1] to confirm equivalence of models:")
-    print(f"     y = {y_check[0]}")
+    print(f"    - true scale:    y = {y_check[0]}")
+
+    # Checking accuracy of solution using normalized values (to avoid numerical errors):
+    x_norm = f.clean(x_norm, SingleInstance=True, normalize=False)  # automatically format the normalized solution
+    y_check_norm = f.evaluate(x_norm, draws=scenarios)
+    print(f"    - normalized:    y = {y_check_norm[0]}")
 
 
 if __name__ == '__main__':
